refactor: refactoring logic due to changes to database
