+++ BPR2-T2/Domain/Dtos/ReservationDtos/AddReservationDto.cs
@@ -11,4 +11,5 @@ public class AddReservationDto
     public DateTime Date { get; set; }
     public string Time { get; set; }
     public int NumOfPeople { get; set; }
+    
 }
