+++ BPR2-T2/Domain/Dtos/RestaurantDtos/CreateRestaurantDto.cs
@@ -1,4 +1,4 @@
-namespace Domain.Dtos;
+namespace Domain.Dtos.RestaurantDtos;
 
 public class CreateRestaurantDto
 {
@@ -8,6 +8,6 @@ public class CreateRestaurantDto
     public string OpenHours { get; set; }
     public string Cuisine { get; set; }
     public string Info { get; set; }
-    public double? Latitude { get; set; }
-    public double? Longitude { get; set; }
+    
+    public int Capacity { get; set; }
 }
