+++ BPR2-T2/Domain/Dtos/RestaurantDtos/UpdateRestaurantDto.cs
@@ -9,5 +9,7 @@ public class UpdateRestaurantDto
     public string OpenHours { get; set; }
     public string Cuisine { get; set; }
     public string Info { get; set; }
+    
+    public int Capacity { get; set; }
     public List<string> ImageUris { get; set; }
 }
